requirements.txt
django==2.2
djangorestframework==3.11.0
PyYAML==5.2
coreapi==2.3.3
Pygments==2.5.2
